pyproject.toml
[project]
name = "cmTools"
version = "0.0.1"
description = "A citation manager for the diSimplex projects which understands our NikolaBase citation system"
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "pyyaml>=6.0",
    "pybtex>=0.24.0",
    "types-PyYAML>=6.0.12.20250915",
]
requires-python = ">=3.10"
license = {text = "Apache2.0"}

[project.scripts]
cmParse    = "cmTools.parser:cli"
cmCollate  = "cmTools.collator:cli"
cmGenerate = "cmTools.generator:cli"
cmScan     = "cmTools.scanner:cli"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.mypy]
# mypy_path = [ "./.venv/lib/python3.13/site-packages/" ]
python_executable = "./.venv/bin/python"

